pyproject.toml
[project]
name = "rs-back"
version = "0.1.1"
description = ""
authors = [
    { name = "zhikh", email = "zhikh.k@gmail.com" },
    { name = "ekirill", email = "kirillv.2016@mail.ru" },
]
readme = "README.md"
requires-python = "^3.12"

dependencies = [
    "django (>=5.2.1)",
    "django-debug-toolbar (>=5.0.1)",
    "drf-spectacular (>=0.28.0)",
    "django-cors-headers (>=4.7.0)",
    "djangorestframework (>=3.16.0)",
    "gunicorn (>=23.0.0)",
    "pillow (>=11.2.1)",
    "python-dotenv (>=1.1.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
